# Conflicts:
#	mRemoteNGTests/Config/Serializers/XmlConnectionNodeSerializerTests.cs
#	mRemoteNGTests/Connection/AbstractConnectionInfoDataTests.cs
#	mRemoteNGTests/app.config
#	mRemoteNGTests/mRemoteNGTests.csproj
#	mRemoteV1/App/Runtime.cs
#	mRemoteV1/Config/Serializers/ConnectionSerializers/Xml/XmlConnectionsDeserializer.cs
#	mRemoteV1/Config/Serializers/CsvConnectionsSerializerMremotengFormat.cs
#	mRemoteV1/Config/Serializers/DataTableDeserializer.cs
#	mRemoteV1/Connection/AbstractConnectionRecord.cs
#	mRemoteV1/Connection/ConnectionInfo.cs
#	mRemoteV1/Connection/Protocol/ProtocolFactory.cs
#	mRemoteV1/UI/Window/ConnectionWindow.cs
#	mRemoteV1/app.config
#	mRemoteV1/mRemoteV1.csproj
#	mRemoteV1/packages.config

+++ mRemoteV1/UI/Window/UpdateWindow.cs
@@ -2,6 +2,7 @@ using System;
 using System.ComponentModel;
 using System.Diagnostics;
 using System.Drawing;
+using System.IO;
 using System.Net;
 using System.Windows.Forms;
 using mRemoteNG.App;
@@ -9,43 +10,53 @@ using mRemoteNG.App.Update;
 using mRemoteNG.Messages;
 using WeifenLuo.WinFormsUI.Docking;
 
-#if !PORTABLE
-using System.IO;
-#endif
-
-
 namespace mRemoteNG.UI.Window
 {
 	public partial class UpdateWindow : BaseWindow
 	{
-#region Public Methods
+        private AppUpdater _appUpdate;
+        private bool _isUpdateDownloadHandlerDeclared;
+
+        #region Public Methods
+        public UpdateWindow() : this(new DockContent())
+	    {
+	    }
+
 		public UpdateWindow(DockContent panel)
 		{
 			WindowType = WindowType.Update;
 			DockPnl = panel;
 			InitializeComponent();
-			Runtime.FontOverride(this);
+            FontOverrider.FontOverride(this);
 		}
-#endregion
+        #endregion
 		
-#region Form Stuff
+        #region Form Stuff
 
 	    private void Update_Load(object sender, EventArgs e)
 		{
-			ApplyLanguage();
+            ApplyTheme();
+            Themes.ThemeManager.getInstance().ThemeChanged += ApplyTheme;
+            ApplyLanguage();
 			CheckForUpdate();
 		}
-				
-		private void ApplyLanguage()
+
+        private new void ApplyTheme()
+        {
+	        if (!Themes.ThemeManager.getInstance().ThemingActive)
+				return;
+
+	        base.ApplyTheme();
+	        pnlUpdate.BackColor = Themes.ThemeManager.getInstance().ActiveTheme.ExtendedPalette.getColor("Dialog_Background");
+	        pnlUpdate.ForeColor = Themes.ThemeManager.getInstance().ActiveTheme.ExtendedPalette.getColor("Dialog_Foreground");
+        }
+
+        private void ApplyLanguage()
 		{
 			Text = Language.strMenuCheckForUpdates;
 			TabText = Language.strMenuCheckForUpdates;
 			btnCheckForUpdate.Text = Language.strCheckForUpdate;
-#if !PORTABLE
-            btnDownload.Text = Language.strDownloadAndInstall;
-#else
-		    btnDownload.Text = Language.strDownloadPortable;
-#endif
+			btnDownload.Text = Runtime.IsPortableEdition ? Language.strDownloadPortable : Language.strDownloadAndInstall;
             lblChangeLogLabel.Text = Language.strLabelChangeLog;
 			lblInstalledVersion.Text = Language.strVersion;
 			lblInstalledVersionLabel.Text = $"{Language.strCurrentVersion}:";
@@ -72,14 +83,9 @@ namespace mRemoteNG.UI.Window
 			}
 			Process.Start(linkUri.ToString());
 		}
-#endregion
+        #endregion
 		
-#region Private Fields
-		private AppUpdater _appUpdate;
-		private bool _isUpdateDownloadHandlerDeclared;
-#endregion
-		
-#region Private Methods
+        #region Private Methods
 		private void CheckForUpdate()
 		{
 			if (_appUpdate == null)
@@ -178,7 +184,7 @@ namespace mRemoteNG.UI.Window
 				lblStatus.Text = Language.strUpdateCheckFailedLabel;
 				lblStatus.ForeColor = Color.OrangeRed;
 						
-				Runtime.MessageCollector.AddExceptionStackTrace(Language.strUpdateCheckCompleteFailed, ex);
+				Runtime.MessageCollector?.AddExceptionStackTrace(Language.strUpdateCheckCompleteFailed, ex);
 			}
 		}
 				
@@ -204,7 +210,7 @@ namespace mRemoteNG.UI.Window
             }
 			catch (Exception ex)
 			{
-				Runtime.MessageCollector.AddExceptionStackTrace(Language.strUpdateGetChangeLogFailed, ex);
+				Runtime.MessageCollector?.AddExceptionStackTrace(Language.strUpdateGetChangeLogFailed, ex);
 			}
 		}
 				
@@ -227,12 +233,12 @@ namespace mRemoteNG.UI.Window
 			}
 			catch (Exception ex)
 			{
-				Runtime.MessageCollector.AddExceptionStackTrace(Language.strUpdateDownloadFailed, ex);
+				Runtime.MessageCollector?.AddExceptionStackTrace(Language.strUpdateDownloadFailed, ex);
 			}
 		}
-#endregion
+        #endregion
 		
-#region Events
+        #region Events
 		private void DownloadUpdateProgressChanged(object sender, DownloadProgressChangedEventArgs e)
 		{
 			prgbDownload.Value = e.ProgressPercentage;
@@ -249,25 +255,27 @@ namespace mRemoteNG.UI.Window
                     return;
 				if (e.Error != null)
 				    throw e.Error;
-#if !PORTABLE
-                if (MessageBox.Show(Language.strUpdateDownloadComplete, Language.strMenuCheckForUpdates, MessageBoxButtons.OKCancel, MessageBoxIcon.Warning) == DialogResult.OK)
+
+				if (Runtime.IsPortableEdition)
+					MessageBox.Show(Language.strUpdatePortableDownloadComplete, Language.strMenuCheckForUpdates, MessageBoxButtons.OK, MessageBoxIcon.Information);
+				else
 				{
-					Shutdown.Quit(_appUpdate.CurrentUpdateInfo.UpdateFilePath);
+					if (MessageBox.Show(Language.strUpdateDownloadComplete, Language.strMenuCheckForUpdates, MessageBoxButtons.OKCancel, MessageBoxIcon.Warning) == DialogResult.OK)
+					{
+						Shutdown.Quit(_appUpdate.CurrentUpdateInfo.UpdateFilePath);
+					}
+					else
+					{
+						File.Delete(_appUpdate.CurrentUpdateInfo.UpdateFilePath);
+					}
 				}
-                else
-				{
-					File.Delete(_appUpdate.CurrentUpdateInfo.UpdateFilePath);
-				}
-#else
-			    MessageBox.Show(Language.strUpdatePortableDownloadComplete, Language.strMenuCheckForUpdates, MessageBoxButtons.OK, MessageBoxIcon.Information);
-#endif
 			}
 			catch (Exception ex)
 			{
-                Runtime.MessageCollector.AddExceptionStackTrace(Language.strUpdateDownloadCompleteFailed, ex);
-                Runtime.MessageCollector.AddMessage(MessageClass.ErrorMsg, ex.Message);
+                Runtime.MessageCollector?.AddExceptionStackTrace(Language.strUpdateDownloadCompleteFailed, ex);
+                Runtime.MessageCollector?.AddMessage(MessageClass.ErrorMsg, ex.Message);
             }
 		}
-#endregion
+        #endregion
 	}
 }
